requirements.txt
streamlit>=1.30.0
PyGithub>=2.0.0
feedparser>=6.0.10
google-generativeai>=0.3.2
python-dotenv>=1.0.0
